refactor: rename js/shapes/atmos → js/scene/atmos
Introduces js/scene/ as the home for complex 3D scene objects,
distinct from js/shapes/ (2D geometry primitives). Updates import
paths in Planet.js, Star.js, and ThreeUI.js.
