+++ js/scene/atmos/AtmospherePrecompute.js
